requirements.txt
Flask==2.0.1
textblob==0.15.3
requests==2.0.5
tweepy==3.10.0
pandas==2.8.2
python-dotenv==0.19.0